@@ -8,7 +8,7 @@ description: >
 
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   analyzer: ^1.7.2
