@@ -38,7 +38,7 @@ class GeneratedFactoryMigrator extends RecursiveAstVisitor
         return;
       }
 
-      yieldPatch('$castFunctionName(', generatedFactory.offset,
+      yieldPatch('$castFunctionName(', generatedFactory!.offset,
           generatedFactory.offset);
       yieldPatch(')', generatedFactory.end, generatedFactory.end);
     }
