@@ -20,7 +20,7 @@ import '../util.dart';
 
 main() {
   group('ReactDomRenderMigrator', () {
-    final testSuggestor = getSuggestorTester(ReactDomRenderMigrator());
+    final Future<void> Function({String expectedOutput, int expectedPatchCount, String input, bool shouldDartfmtOutput, bool testIdempotency, void Function(String) validateContents}) testSuggestor = getSuggestorTester(ReactDomRenderMigrator() as Stream<Patch> Function(FileContext));
 
     test('empty file', () async {
       await testSuggestor(expectedPatchCount: 0, input: '');
